Add validation successfully

+++ src/main/java/com/codegym/employee/repository/EmployeeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface EmployeeRepository extends PagingAndSortingRepository<Employee, Integer> {
+public interface EmployeeRepository extends PagingAndSortingRepository<Employee, Long> {
     Page<Employee> findAllByFullNameContaining(String fullname, Pageable pageable);
 }
